backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
email-validator
pydantic
aiofiles
jinja2
python-multipart
cryptography